# some basic dag optimizations