--- a/dag/set.go
+++ b/dag/set.go
@@ -1,14 +1,7 @@
 package dag
 
-import (
-	"sync"
-)
-
 // Set is a set data structure.
-type Set struct {
-	m    map[interface{}]interface{}
-	once sync.Once
-}
+type Set map[interface{}]interface{}
 
 // Hashable is the interface used by set to get the hash code of a value.
 // If this isn't given, then the value of the item being added to the set
@@ -27,32 +20,29 @@ func hashcode(v interface{}) interface{} {
 }
 
 // Add adds an item to the set
-func (s *Set) Add(v interface{}) {
-	s.once.Do(s.init)
-	s.m[hashcode(v)] = v
+func (s Set) Add(v interface{}) {
+	s[hashcode(v)] = v
 }
 
 // Delete removes an item from the set.
-func (s *Set) Delete(v interface{}) {
-	s.once.Do(s.init)
-	delete(s.m, hashcode(v))
+func (s Set) Delete(v interface{}) {
+	delete(s, hashcode(v))
 }
 
 // Include returns true/false of whether a value is in the set.
-func (s *Set) Include(v interface{}) bool {
-	s.once.Do(s.init)
-	_, ok := s.m[hashcode(v)]
+func (s Set) Include(v interface{}) bool {
+	_, ok := s[hashcode(v)]
 	return ok
 }
 
 // Intersection computes the set intersection with other.
-func (s *Set) Intersection(other *Set) *Set {
-	result := new(Set)
+func (s Set) Intersection(other Set) Set {
+	result := make(Set)
 	if s == nil {
 		return result
 	}
 	if other != nil {
-		for _, v := range s.m {
+		for _, v := range s {
 			if other.Include(v) {
 				result.Add(v)
 			}
@@ -64,13 +54,13 @@ func (s *Set) Intersection(other *Set) *Set {
 
 // Difference returns a set with the elements that s has but
 // other doesn't.
-func (s *Set) Difference(other *Set) *Set {
-	result := new(Set)
+func (s Set) Difference(other Set) Set {
+	result := make(Set)
 	if s != nil {
-		for k, v := range s.m {
+		for k, v := range s {
 			var ok bool
 			if other != nil {
-				_, ok = other.m[k]
+				_, ok = other[k]
 			}
 			if !ok {
 				result.Add(v)
@@ -83,10 +73,10 @@ func (s *Set) Difference(other *Set) *Set {
 
 // Filter returns a set that contains the elements from the receiver
 // where the given callback returns true.
-func (s *Set) Filter(cb func(interface{}) bool) *Set {
-	result := new(Set)
+func (s Set) Filter(cb func(interface{}) bool) Set {
+	result := make(Set)
 
-	for _, v := range s.m {
+	for _, v := range s {
 		if cb(v) {
 			result.Add(v)
 		}
@@ -96,28 +86,20 @@ func (s *Set) Filter(cb func(interface{}) bool) *Set {
 }
 
 // Len is the number of items in the set.
-func (s *Set) Len() int {
-	if s == nil {
-		return 0
-	}
-
-	return len(s.m)
+func (s Set) Len() int {
+	return len(s)
 }
 
 // List returns the list of set elements.
-func (s *Set) List() []interface{} {
+func (s Set) List() []interface{} {
 	if s == nil {
 		return nil
 	}
 
-	r := make([]interface{}, 0, len(s.m))
-	for _, v := range s.m {
+	r := make([]interface{}, 0, len(s))
+	for _, v := range s {
 		r = append(r, v)
 	}
 
 	return r
-}
-
-func (s *Set) init() {
-	s.m = make(map[interface{}]interface{})
 }