--- a/dag/tarjan.go
+++ b/dag/tarjan.go
@@ -24,7 +24,7 @@ func stronglyConnected(acct *sccAcct, g *Graph, v Vertex) int {
 	index := acct.visit(v)
 	minIdx := index
 
-	for _, raw := range g.DownEdges(v).List() {
+	for _, raw := range g.DownEdges(v) {
 		target := raw.(Vertex)
 		targetIdx := acct.VertexIndex[target]