--- a/dag/walk.go
+++ b/dag/walk.go
@@ -174,7 +174,7 @@ func (w *Walker) Update(g *AcyclicGraph) {
 	oldVerts := w.vertices.Difference(v)
 
 	// Add the new vertices
-	for _, raw := range newVerts.List() {
+	for _, raw := range newVerts {
 		v := raw.(Vertex)
 
 		// Add to the waitgroup so our walk is not done until everything finishes
@@ -196,7 +196,7 @@ func (w *Walker) Update(g *AcyclicGraph) {
 	}
 
 	// Remove the old vertices
-	for _, raw := range oldVerts.List() {
+	for _, raw := range oldVerts {
 		v := raw.(Vertex)
 
 		// Get the vertex info so we can cancel it
@@ -219,7 +219,7 @@ func (w *Walker) Update(g *AcyclicGraph) {
 
 	// Add the new edges
 	changedDeps := make(Set)
-	for _, raw := range newEdges.List() {
+	for _, raw := range newEdges {
 		edge := raw.(Edge)
 		waiter, dep := w.edgeParts(edge)
 
@@ -249,8 +249,8 @@ func (w *Walker) Update(g *AcyclicGraph) {
 		w.edges.Add(raw)
 	}
 
-	// Process reoved edges
-	for _, raw := range oldEdges.List() {
+	// Process removed edges
+	for _, raw := range oldEdges {
 		edge := raw.(Edge)
 		waiter, dep := w.edgeParts(edge)
 
@@ -275,7 +275,7 @@ func (w *Walker) Update(g *AcyclicGraph) {
 
 	// For each vertex with changed dependencies, we need to kick off
 	// a new waiter and notify the vertex of the changes.
-	for _, raw := range changedDeps.List() {
+	for _, raw := range changedDeps {
 		v := raw.(Vertex)
 		info, ok := w.vertexMap[v]
 		if !ok {
@@ -320,7 +320,7 @@ func (w *Walker) Update(g *AcyclicGraph) {
 
 	// Start all the new vertices. We do this at the end so that all
 	// the edge waiters and changes are setup above.
-	for _, raw := range newVerts.List() {
+	for _, raw := range newVerts {
 		v := raw.(Vertex)
 		go w.walkVertex(v, w.vertexMap[v])
 	}