--- a/terraform/graph.go
+++ b/terraform/graph.go
@@ -20,11 +20,6 @@ type Graph struct {
 
 	// Path is the path in the module tree that this Graph represents.
 	Path addrs.ModuleInstance
-
-	// debugName is a name for reference in the debug output. This is usually
-	// to indicate what topmost builder was, and if this graph is a shadow or
-	// not.
-	debugName string
 }
 
 func (g *Graph) DirectedGraph() dag.Grapher {
@@ -43,19 +38,10 @@ func (g *Graph) walk(walker GraphWalker) tfdiags.Diagnostics {
 	ctx := walker.EnterPath(g.Path)
 	defer walker.ExitPath(g.Path)
 
-	// Get the path for logs
-	path := ctx.Path().String()
-
-	debugName := "walk-graph.json"
-	if g.debugName != "" {
-		debugName = g.debugName + "-" + debugName
-	}
-
 	// Walk the graph.
 	var walkFn dag.WalkFunc
 	walkFn = func(v dag.Vertex) (diags tfdiags.Diagnostics) {
 		log.Printf("[TRACE] vertex %q: starting visit (%T)", dag.VertexName(v), v)
-		g.DebugVisitInfo(v, g.debugName)
 
 		defer func() {
 			log.Printf("[TRACE] vertex %q: visit complete", dag.VertexName(v))
@@ -84,8 +70,6 @@ func (g *Graph) walk(walker GraphWalker) tfdiags.Diagnostics {
 			// then callback with the output.
 			log.Printf("[TRACE] vertex %q: evaluating", dag.VertexName(v))
 
-			g.DebugVertexInfo(v, fmt.Sprintf("evaluating %T(%s)", v, path))
-
 			tree = walker.EnterEvalTree(v, tree)
 			output, err := Eval(tree, vertexCtx)
 			diags = diags.Append(walker.ExitEvalTree(v, output, err))
@@ -97,8 +81,6 @@ func (g *Graph) walk(walker GraphWalker) tfdiags.Diagnostics {
 		// If the node is dynamically expanded, then expand it
 		if ev, ok := v.(GraphNodeDynamicExpandable); ok {
 			log.Printf("[TRACE] vertex %q: expanding dynamic subgraph", dag.VertexName(v))
-
-			g.DebugVertexInfo(v, fmt.Sprintf("expanding %T(%s)", v, path))
 
 			g, err := ev.DynamicExpand(vertexCtx)
 			if err != nil {
@@ -123,8 +105,6 @@ func (g *Graph) walk(walker GraphWalker) tfdiags.Diagnostics {
 		// If the node has a subgraph, then walk the subgraph
 		if sn, ok := v.(GraphNodeSubgraph); ok {
 			log.Printf("[TRACE] vertex %q: entering static subgraph", dag.VertexName(v))
-
-			g.DebugVertexInfo(v, fmt.Sprintf("subgraph: %T(%s)", v, path))
 
 			subDiags := sn.Subgraph().(*Graph).walk(walker)
 			if subDiags.HasErrors() {