--- a/terraform/transform_destroy_edge.go
+++ b/terraform/transform_destroy_edge.go
@@ -178,12 +178,11 @@ func (t *DestroyEdgeTransformer) pruneResources(g *Graph) error {
 		}
 
 		// if there are only destroy dependencies, we don't need this node
-		des, err := g.Descendents(n)
+		descendents, err := g.Descendents(n)
 		if err != nil {
 			return err
 		}
 
-		descendents := des.List()
 		nonDestroyInstanceFound := false
 		for _, v := range descendents {
 			if _, ok := v.(*NodeApplyableResourceInstance); ok {
@@ -197,8 +196,8 @@ func (t *DestroyEdgeTransformer) pruneResources(g *Graph) error {
 		}
 
 		// connect all the through-edges, then delete the node
-		for _, d := range g.DownEdges(n).List() {
-			for _, u := range g.UpEdges(n).List() {
+		for _, d := range g.DownEdges(n) {
+			for _, u := range g.UpEdges(n) {
 				g.Connect(dag.BasicEdge(u, d))
 			}
 		}