… command/debug_json2dot.go
… command/debug_json2dot_test.go
… dag/dag_test.go
… dag/graph_test.go
… dag/marshal.go
… dag/marshal_test.go
… dag/set_test.go
… terraform/transform_output.go
… terraform/transform_provider.go
… terraform/transform_reference.go
… terraform/transform_targets.go