@jbardin 2020-01-08
Added a fairly coarse benchmark, comprised of the 2 most expensive operations usually done by core -- find all node dependencies, and reduce the graph. While not prefect, it provides some numbers to start with.

```
benchmark          old ns/op       new ns/op      delta
BenchmarkDAG-4     28129990770     4288588518     -84.75%

benchmark          old allocs     new allocs     delta
BenchmarkDAG-4     247964230      22401398       -90.97%

benchmark          old bytes      new bytes      delta
BenchmarkDAG-4     2913358152     1395783960     -52.09%
```